+++ gui/extension/CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Additions
 
-- 
+-
 
 ### Fixes
 
-- 
+- BUG#37716544 Conflicts when updating a REST Document with the TypeScript SDK
 
 ## Changes in 1.19.14+9.3.1